+++ debug/example/cachememory/page1.class.js
@@ -1,3 +1,4 @@
+import AsyncTask from "../../flexvue/core/asynctask.class.js";
 class ComponentActivity {
   constructor() {
     this.TAG = 'page1';
@@ -5,9 +6,17 @@ class ComponentActivity {
   doList(params = {}) {
     var _a;
     Log.v(this.TAG, 'doList', params);
-    // 캐시메모리
-    const cachedData = (_a = window.cacheMemory._get('/test/object')) !== null && _a !== void 0 ? _a : '캐시데이터가 없거나 만료됨';
-    Log.e('cache', '/test/object', cachedData);
+    // promise.all 사용
+    Promise.all([
+    (_a = window.cacheMemory._getAsync('/test/object')) !== null && _a !== void 0 ? _a : new AsyncTask().doImport(new URL('../js/values/arrays.js', import.meta.url).href),
+    new AsyncTask().doImport(new URL('../js/values/strings.js', import.meta.url).href)]
+    ).
+    then((data) => {
+      Log.v('Promise.all', data);
+    }).
+    catch((e) => {
+      Log.e(e);
+    });
   }
   doPost(params = {}) {
     Log.v(this.TAG, 'doPost', params);