+++ debug/example/cachestorage/cachestorage.js
@@ -1,6 +1,7 @@
 import { CacheLocalStorage } from "../../flexvue/core/caches.class.js";
+import AsyncTask from "../../flexvue/core/asynctask.class.js";
 const onReady = () => {
-  var _a, _b, _c;
+  var _a, _b, _c, _d;
   new App();
   Log.i(App.browser, App.version, App.os, App.lang);
   // 캐시 클래스 글로벌 선언
@@ -22,6 +23,17 @@ const onReady = () => {
   window.cacheStorage._getAsync('mc_array').
   then((data) => {
     Log.d('_getAsync', data);
+  });
+  // promise.all 사용
+  Promise.all([
+  (_d = window.cacheStorage._getAsync('mc_array')) !== null && _d !== void 0 ? _d : new AsyncTask().doImport(new URL('../js/values/arrays.js', import.meta.url).href),
+  new AsyncTask().doImport(new URL('../js/values/strings.js', import.meta.url).href)]
+  ).
+  then((data) => {
+    Log.d(data);
+  }).
+  catch((e) => {
+    Log.e(e);
   });
   // 캐시 삭제
   // window.cacheStorage._delete('mc');