+++ debug/plugins/apexcharts/src/svgjs/svg.js
@@ -1984,6 +1984,24 @@
 
 
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
         // functionToCall: [list of morphable objects]
         // e.g. move: [SVG.Number, SVG.Number]
       };this.attrs = {
@@ -2109,6 +2127,15 @@
         if (element instanceof SVG.Element) {var box; // yes this is ugly, but Firefox can be a pain when it comes to elements that are not yet rendered
           try {if (!document.documentElement.contains) {// This is IE - it does not support contains() for top-level SVGs
               var topParent = element.node;while (topParent.parentNode) {topParent = topParent.parentNode;}if (topParent != document) throw new Error('Element not in the dom');} else {
+
+
+
+
+
+
+
+
+
 
 
 
@@ -3436,37 +3463,10 @@
   function arrayToMatrix(a) {return { a: a[0], b: a[1], c: a[2], d: a[3], e: a[4], f: a[5] };} // Parse matrix if required
   function parseMatrix(matrix) {if (!(matrix instanceof SVG.Matrix)) {matrix = new SVG.Matrix(matrix);}return matrix;} // Add centre point to transform object
   function ensureCentre(o, target) {o.cx = o.cx == null ? target.bbox().cx : o.cx;o.cy = o.cy == null ? target.bbox().cy : o.cy;} // PathArray Helpers
-  function arrayToString(a) {for (var i = 0, il = a.length, s = ''; i < il; i++) {s += a[i][0];if (a[i][1] != null) {s += a[i][1];if (a[i][2] != null) {s += ' ';s += a[i][2];if (a[i][3] != null) {s += ' ';s += a[i][3];s += ' ';s += a[i][4];if (a[i][5] != null) {s += ' ';s += a[i][5];s += ' ';s += a[i][6];if (a[i][7] != null) {s += ' ';s += a[i][7];}
-            }
-          }
-        }
-      }
-    }
-    return s + ' ';
-  }
-
-  // Deep new id assignment
-  function assignNewId(node) {
-    // do the same for SVG child nodes as well
-    for (var i = node.childNodes.length - 1; i >= 0; i--) {
-      if (node.childNodes[i] instanceof window.SVGElement) {
-        assignNewId(node.childNodes[i]);
-      }
-    }
-    return SVG.adopt(node).id(SVG.eid(node.nodeName));
-  }
-
-  // Add more bounding box properties
-  function fullBox(b) {
-    if (b.x == null) {
-      b.x = 0;
-      b.y = 0;
-      b.width = 0;
-      b.height = 0;
-    }
-    b.w = b.width;
-    b.h = b.height;
-    b.x2 = b.x + b.width;
+  function arrayToString(a) {for (var i = 0, il = a.length, s = ''; i < il; i++) {s += a[i][0];if (a[i][1] != null) {s += a[i][1];if (a[i][2] != null) {s += ' ';s += a[i][2];if (a[i][3] != null) {s += ' ';s += a[i][3];s += ' ';s += a[i][4];if (a[i][5] != null) {s += ' ';s += a[i][5];s += ' ';s += a[i][6];if (a[i][7] != null) {s += ' ';s += a[i][7];}}}}}}return s + ' ';} // Deep new id assignment
+  function assignNewId(node) {// do the same for SVG child nodes as well
+    for (var i = node.childNodes.length - 1; i >= 0; i--) {if (node.childNodes[i] instanceof window.SVGElement) {assignNewId(node.childNodes[i]);}}return SVG.adopt(node).id(SVG.eid(node.nodeName));} // Add more bounding box properties
+  function fullBox(b) {if (b.x == null) {b.x = 0;b.y = 0;b.width = 0;b.height = 0;}b.w = b.width;b.h = b.height;b.x2 = b.x + b.width;
     b.y2 = b.y + b.height;
     b.cx = b.x + b.width / 2;
     b.cy = b.y + b.height / 2;