+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [3.0]
 ### - 2024-04-13
 - CacheMemory,CacheLocalStorage class 비동기 메소드 _getAsync 추가 및 _set 객체object 리턴으로 변경
+- CacheMemory,CacheLocalStorage 비동기 메소드 : Promise 비동기 지원
 
 ### - 2024-04-12
 - RecyclerView 실시간 스크롤 위치 확인 기능 추가